requirements.txt
praw==7.7.1
pyttsx3==2.90
moviepy==1.0.3
pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1